+++ SW.Tools/Helpers/DotNetUtilitiesCustom.cs
