+++ SW.Tools/Helpers/Xml.cs
@@ -1,4 +1,6 @@
-﻿using System.Xml;
+﻿using System.IO;
+using System.Text;
+using System.Xml;
 
 namespace SW.Tools.Helpers
 {
